+++ Rebus/Workers/ThreadPoolBased/DefaultSyncBackoffStrategy.cs
@@ -26,11 +26,31 @@ namespace Rebus.Workers.ThreadPoolBased
             }
         }
 
-        /// <summary>
-        /// Executes the next wait operation by blocking the thread, possibly advancing the wait cursor to a different wait time for the next time.
-        /// This function is called each time no message was received.
-        /// </summary>
+        /// <inheritdoc />
         public void Wait()
+        {
+            InnerWait();
+        }
+
+        /// <inheritdoc />
+        public void WaitNoMessage()
+        {
+            InnerWait();
+        }
+
+        /// <inheritdoc />
+        public void WaitError()
+        {
+            Thread.Sleep(TimeSpan.FromSeconds(5));
+        }
+
+        /// <inheritdoc />
+        public void Reset()
+        {
+            Interlocked.Exchange(ref _waitTimeTicks, 0);
+        }
+
+        void InnerWait()
         {
             var waitedSinceTicks = Interlocked.Read(ref _waitTimeTicks);
 
@@ -41,28 +61,12 @@ namespace Rebus.Workers.ThreadPoolBased
             }
 
             var waitDurationTicks = DateTime.UtcNow.Ticks - waitedSinceTicks;
-            var totalSecondsIdle = (int)TimeSpan.FromTicks(waitDurationTicks).TotalSeconds;
+            var totalSecondsIdle = (int) TimeSpan.FromTicks(waitDurationTicks).TotalSeconds;
             var waitTimeIndex = Math.Min(totalSecondsIdle, _backoffTimes.Length - 1);
 
             var backoffTime = _backoffTimes[waitTimeIndex];
 
             Thread.Sleep(backoffTime);
-        }
-
-        /// <summary>
-        /// Blocks the thread for a (most likely longer) while, when an error has occurred
-        /// </summary>
-        public void WaitError()
-        {
-            Thread.Sleep(TimeSpan.FromSeconds(5));
-        }
-
-        /// <summary>
-        /// Resets the strategy. Is called whenever a message was received.
-        /// </summary>
-        public void Reset()
-        {
-            Interlocked.Exchange(ref _waitTimeTicks, 0);
         }
     }
 }