+++ CHANGELOG.md
@@ -1080,6 +1080,11 @@
 
 * Test release
 
+## 2.0.0-b03
+
+* Add separate wait methods to `ISyncBackoffStrategy` in order to differentiate between waiting because no more parallel operations are allowed, and waiting because no message was received
+* Add `deliveryTimeoutSeconds` paramater to `Deliver` method of `SagaFixture` in order to allow for not timing out when e.g. step-debugging saga message processing
+
 ---
 
 [AndreaCuneo]: https://github.com/AndreaCuneo